@@ -11,14 +11,14 @@ import 'package:args/command_runner.dart';
 import 'package:colorize/colorize.dart';
 import 'package:file/file.dart';
 import 'package:git/git.dart';
-import 'package:meta/meta.dart';
+import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 import 'package:yaml/yaml.dart';
 
 /// The signature for a print handler for commands that allow overriding the
 /// print destination.
-typedef Print = void Function(Object object);
+typedef Print = void Function(Object? object);
 
 /// Key for windows platform.
 const String kWindows = 'windows';
@@ -49,7 +49,7 @@ const String kEnableExperiment = 'enable-experiment';
 
 /// Returns whether the given directory contains a Flutter package.
 bool isFlutterPackage(FileSystemEntity entity, FileSystem fileSystem) {
-  if (entity == null || entity is! Directory) {
+  if (entity is! Directory) {
     return false;
   }
 
@@ -58,7 +58,7 @@ bool isFlutterPackage(FileSystemEntity entity, FileSystem fileSystem) {
         fileSystem.file(p.join(entity.path, 'pubspec.yaml'));
     final YamlMap pubspecYaml =
         loadYaml(pubspecFile.readAsStringSync()) as YamlMap;
-    final YamlMap dependencies = pubspecYaml['dependencies'] as YamlMap;
+    final YamlMap? dependencies = pubspecYaml['dependencies'] as YamlMap?;
     if (dependencies == null) {
       return false;
     }
@@ -86,7 +86,7 @@ bool pluginSupportsPlatform(
       platform == kMacos ||
       platform == kWindows ||
       platform == kLinux);
-  if (entity == null || entity is! Directory) {
+  if (entity is! Directory) {
     return false;
   }
 
@@ -95,15 +95,15 @@ bool pluginSupportsPlatform(
         fileSystem.file(p.join(entity.path, 'pubspec.yaml'));
     final YamlMap pubspecYaml =
         loadYaml(pubspecFile.readAsStringSync()) as YamlMap;
-    final YamlMap flutterSection = pubspecYaml['flutter'] as YamlMap;
+    final YamlMap? flutterSection = pubspecYaml['flutter'] as YamlMap?;
     if (flutterSection == null) {
       return false;
     }
-    final YamlMap pluginSection = flutterSection['plugin'] as YamlMap;
+    final YamlMap? pluginSection = flutterSection['plugin'] as YamlMap?;
     if (pluginSection == null) {
       return false;
     }
-    final YamlMap platforms = pluginSection['platforms'] as YamlMap;
+    final YamlMap? platforms = pluginSection['platforms'] as YamlMap?;
     if (platforms == null) {
       // Legacy plugin specs are assumed to support iOS and Android.
       if (!pluginSection.containsKey('platforms')) {
@@ -150,7 +150,7 @@ bool isLinuxPlugin(FileSystemEntity entity, FileSystem fileSystem) {
 }
 
 /// Throws a [ToolExit] with `exitCode` and log the `errorMessage` in red.
-void printErrorAndExit({@required String errorMessage, int exitCode = 1}) {
+void printErrorAndExit({required String errorMessage, int exitCode = 1}) {
   final Colorize redError = Colorize(errorMessage)..red();
   print(redError);
   throw ToolExit(exitCode);
@@ -203,6 +203,7 @@ abstract class PluginCommand extends Command<void> {
     argParser.addFlag(_runOnChangedPackagesArg,
         help: 'Run the command on changed packages/plugins.\n'
             'If the $_pluginsArg is specified, this flag is ignored.\n'
+            'If no plugins have changed, the command runs on all plugins.\n'
             'The packages excluded with $_excludeArg is also excluded even if changed.\n'
             'See $_kBaseSha if a custom base is needed to determine the diff.');
     argParser.addOption(_kBaseSha,
@@ -234,17 +235,17 @@ abstract class PluginCommand extends Command<void> {
   /// The git directory to use. By default it uses the parent directory.
   ///
   /// This can be mocked for testing.
-  final GitDir gitDir;
+  final GitDir? gitDir;
 
-  int _shardIndex;
-  int _shardCount;
+  int? _shardIndex;
+  int? _shardCount;
 
   /// The shard of the overall command execution that this instance should run.
   int get shardIndex {
     if (_shardIndex == null) {
       _checkSharding();
     }
-    return _shardIndex;
+    return _shardIndex!;
   }
 
   /// The number of shards this command is divided into.
@@ -252,12 +253,27 @@ abstract class PluginCommand extends Command<void> {
     if (_shardCount == null) {
       _checkSharding();
     }
-    return _shardCount;
+    return _shardCount!;
+  }
+
+  /// Convenience accessor for boolean arguments.
+  bool getBoolArg(String key) {
+    return (argResults![key] as bool?) ?? false;
+  }
+
+  /// Convenience accessor for String arguments.
+  String getStringArg(String key) {
+    return (argResults![key] as String?) ?? '';
+  }
+
+  /// Convenience accessor for List<String> arguments.
+  List<String> getStringListArg(String key) {
+    return (argResults![key] as List<String>?) ?? <String>[];
   }
 
   void _checkSharding() {
-    final int shardIndex = int.tryParse(argResults[_shardIndexArg] as String);
-    final int shardCount = int.tryParse(argResults[_shardCountArg] as String);
+    final int? shardIndex = int.tryParse(getStringArg(_shardIndexArg));
+    final int? shardCount = int.tryParse(getStringArg(_shardCountArg));
     if (shardIndex == null) {
       usageException('$_shardIndexArg must be an integer');
     }
@@ -300,8 +316,8 @@ abstract class PluginCommand extends Command<void> {
   /// Returns the root Dart package folders of the plugins involved in this
   /// command execution, assuming there is only one shard.
   ///
-  /// Plugin packages can exist in one of two places relative to the packages
-  /// directory.
+  /// Plugin packages can exist in the following places relative to the packages
+  /// directory:
   ///
   /// 1. As a Dart package in a directory which is a direct child of the
   ///    packages directory. This is a plugin where all of the implementations
@@ -311,44 +327,54 @@ abstract class PluginCommand extends Command<void> {
   ///    packages which implement a single plugin. This directory contains a
   ///    "client library" package, which declares the API for the plugin, as
   ///    well as one or more platform-specific implementations.
+  /// 3./4. Either of the above, but in a third_party/packages/ directory that
+  ///    is a sibling of the packages directory. This is used for a small number
+  ///    of packages in the flutter/packages repository.
   Stream<Directory> _getAllPlugins() async* {
-    Set<String> plugins =
-        Set<String>.from(argResults[_pluginsArg] as List<String>);
+    Set<String> plugins = Set<String>.from(getStringListArg(_pluginsArg));
     final Set<String> excludedPlugins =
-        Set<String>.from(argResults[_excludeArg] as List<String>);
-    final bool runOnChangedPackages =
-        argResults[_runOnChangedPackagesArg] as bool;
+        Set<String>.from(getStringListArg(_excludeArg));
+    final bool runOnChangedPackages = getBoolArg(_runOnChangedPackagesArg);
     if (plugins.isEmpty && runOnChangedPackages) {
       plugins = await _getChangedPackages();
     }
 
-    await for (final FileSystemEntity entity
-        in packagesDir.list(followLinks: false)) {
-      // A top-level Dart package is a plugin package.
-      if (_isDartPackage(entity)) {
-        if (!excludedPlugins.contains(entity.basename) &&
-            (plugins.isEmpty || plugins.contains(p.basename(entity.path)))) {
-          yield entity as Directory;
-        }
-      } else if (entity is Directory) {
-        // Look for Dart packages under this top-level directory.
-        await for (final FileSystemEntity subdir
-            in entity.list(followLinks: false)) {
-          if (_isDartPackage(subdir)) {
-            // If --plugin=my_plugin is passed, then match all federated
-            // plugins under 'my_plugin'. Also match if the exact plugin is
-            // passed.
-            final String relativePath =
-                p.relative(subdir.path, from: packagesDir.path);
-            final String packageName = p.basename(subdir.path);
-            final String basenamePath = p.basename(entity.path);
-            if (!excludedPlugins.contains(basenamePath) &&
-                !excludedPlugins.contains(packageName) &&
-                !excludedPlugins.contains(relativePath) &&
-                (plugins.isEmpty ||
-                    plugins.contains(relativePath) ||
-                    plugins.contains(basenamePath))) {
-              yield subdir as Directory;
+    final Directory thirdPartyPackagesDirectory = packagesDir.parent
+        .childDirectory('third_party')
+        .childDirectory('packages');
+
+    for (final Directory dir in <Directory>[
+      packagesDir,
+      if (thirdPartyPackagesDirectory.existsSync()) thirdPartyPackagesDirectory,
+    ]) {
+      await for (final FileSystemEntity entity
+          in dir.list(followLinks: false)) {
+        // A top-level Dart package is a plugin package.
+        if (_isDartPackage(entity)) {
+          if (!excludedPlugins.contains(entity.basename) &&
+              (plugins.isEmpty || plugins.contains(p.basename(entity.path)))) {
+            yield entity as Directory;
+          }
+        } else if (entity is Directory) {
+          // Look for Dart packages under this top-level directory.
+          await for (final FileSystemEntity subdir
+              in entity.list(followLinks: false)) {
+            if (_isDartPackage(subdir)) {
+              // If --plugin=my_plugin is passed, then match all federated
+              // plugins under 'my_plugin'. Also match if the exact plugin is
+              // passed.
+              final String relativePath =
+                  p.relative(subdir.path, from: dir.path);
+              final String packageName = p.basename(subdir.path);
+              final String basenamePath = p.basename(entity.path);
+              if (!excludedPlugins.contains(basenamePath) &&
+                  !excludedPlugins.contains(packageName) &&
+                  !excludedPlugins.contains(relativePath) &&
+                  (plugins.isEmpty ||
+                      plugins.contains(relativePath) ||
+                      plugins.contains(basenamePath))) {
+                yield subdir as Directory;
+              }
             }
           }
         }
@@ -415,9 +441,9 @@ abstract class PluginCommand extends Command<void> {
   /// Throws tool exit if [gitDir] nor root directory is a git directory.
   Future<GitVersionFinder> retrieveVersionFinder() async {
     final String rootDir = packagesDir.parent.absolute.path;
-    final String baseSha = argResults[_kBaseSha] as String;
+    final String baseSha = getStringArg(_kBaseSha);
 
-    GitDir baseGitDir = gitDir;
+    GitDir? baseGitDir = gitDir;
     if (baseGitDir == null) {
       if (!await GitDir.isGitDir(rootDir)) {
         printErrorAndExit(
@@ -449,8 +475,9 @@ abstract class PluginCommand extends Command<void> {
     if (packages.isNotEmpty) {
       final String changedPackages = packages.join(',');
       print(changedPackages);
+    } else {
+      print('No changed packages.');
     }
-    print('No changed packages.');
     return packages;
   }
 }
@@ -475,7 +502,7 @@ class ProcessRunner {
   Future<int> runAndStream(
     String executable,
     List<String> args, {
-    Directory workingDir,
+    Directory? workingDir,
     bool exitOnError = false,
   }) async {
     print(
@@ -500,17 +527,33 @@ class ProcessRunner {
   ///
   /// If [exitOnError] is set to `true`, then this will throw an error if
   /// the [executable] terminates with a non-zero exit code.
+  /// Defaults to `false`.
+  ///
+  /// If [logOnError] is set to `true`, it will print a formatted message about the error.
+  /// Defaults to `false`
   ///
   /// Returns the [io.ProcessResult] of the [executable].
   Future<io.ProcessResult> run(String executable, List<String> args,
-      {Directory workingDir,
+      {Directory? workingDir,
       bool exitOnError = false,
+      bool logOnError = false,
       Encoding stdoutEncoding = io.systemEncoding,
       Encoding stderrEncoding = io.systemEncoding}) async {
-    return io.Process.run(executable, args,
+    final io.ProcessResult result = await io.Process.run(executable, args,
         workingDirectory: workingDir?.path,
         stdoutEncoding: stdoutEncoding,
         stderrEncoding: stderrEncoding);
+    if (result.exitCode != 0) {
+      if (logOnError) {
+        final String error =
+            _getErrorString(executable, args, workingDir: workingDir);
+        print('$error Stderr:\n${result.stdout}');
+      }
+      if (exitOnError) {
+        throw ToolExit(result.exitCode);
+      }
+    }
+    return result;
   }
 
   /// Starts the [executable] with [args].
@@ -519,43 +562,110 @@ class ProcessRunner {
   /// passing [workingDir].
   ///
   /// Returns the started [io.Process].
-  Future<io.Process> start(String executable, List<String> args,
-      {Directory workingDirectory}) async {
+  Future<io.Process?> start(String executable, List<String> args,
+      {Directory? workingDirectory}) async {
     final io.Process process = await io.Process.start(executable, args,
         workingDirectory: workingDirectory?.path);
     return process;
   }
 
-  /// Run the [executable] with [args], throwing an error on non-zero exit code.
-  ///
-  /// Unlike [runAndStream], this does not stream the process output to stdout.
-  /// It also unconditionally throws an error on a non-zero exit code.
-  ///
-  /// The current working directory of [executable] can be overridden by
-  /// passing [workingDir].
-  ///
-  /// Returns the [io.ProcessResult] of running the [executable].
-  Future<io.ProcessResult> runAndExitOnError(
-    String executable,
-    List<String> args, {
-    Directory workingDir,
-  }) async {
-    final io.ProcessResult result = await io.Process.run(executable, args,
-        workingDirectory: workingDir?.path);
-    if (result.exitCode != 0) {
-      final String error =
-          _getErrorString(executable, args, workingDir: workingDir);
-      print('$error Stderr:\n${result.stdout}');
-      throw ToolExit(result.exitCode);
-    }
-    return result;
-  }
-
   String _getErrorString(String executable, List<String> args,
-      {Directory workingDir}) {
+      {Directory? workingDir}) {
     final String workdir = workingDir == null ? '' : ' in ${workingDir.path}';
     return 'ERROR: Unable to execute "$executable ${args.join(' ')}"$workdir.';
   }
+}
+
+/// Finding version of [package] that is published on pub.
+class PubVersionFinder {
+  /// Constructor.
+  ///
+  /// Note: you should manually close the [httpClient] when done using the finder.
+  PubVersionFinder({this.pubHost = defaultPubHost, required this.httpClient});
+
+  /// The default pub host to use.
+  static const String defaultPubHost = 'https://pub.dev';
+
+  /// The pub host url, defaults to `https://pub.dev`.
+  final String pubHost;
+
+  /// The http client.
+  ///
+  /// You should manually close this client when done using this finder.
+  final http.Client httpClient;
+
+  /// Get the package version on pub.
+  Future<PubVersionFinderResponse> getPackageVersion(
+      {required String package}) async {
+    assert(package.isNotEmpty);
+    final Uri pubHostUri = Uri.parse(pubHost);
+    final Uri url = pubHostUri.replace(path: '/packages/$package.json');
+    final http.Response response = await httpClient.get(url);
+
+    if (response.statusCode == 404) {
+      return PubVersionFinderResponse(
+          versions: null,
+          result: PubVersionFinderResult.noPackageFound,
+          httpResponse: response);
+    } else if (response.statusCode != 200) {
+      return PubVersionFinderResponse(
+          versions: null,
+          result: PubVersionFinderResult.fail,
+          httpResponse: response);
+    }
+    final List<Version> versions =
+        (json.decode(response.body)['versions'] as List<dynamic>)
+            .map<Version>((final dynamic versionString) =>
+                Version.parse(versionString as String))
+            .toList();
+
+    return PubVersionFinderResponse(
+        versions: versions,
+        result: PubVersionFinderResult.success,
+        httpResponse: response);
+  }
+}
+
+/// Represents a response for [PubVersionFinder].
+class PubVersionFinderResponse {
+  /// Constructor.
+  PubVersionFinderResponse({this.versions, this.result, this.httpResponse}) {
+    if (versions != null && versions!.isNotEmpty) {
+      versions!.sort((Version a, Version b) {
+        // TODO(cyanglaz): Think about how to handle pre-release version with [Version.prioritize].
+        // https://github.com/flutter/flutter/issues/82222
+        return b.compareTo(a);
+      });
+    }
+  }
+
+  /// The versions found in [PubVersionFinder].
+  ///
+  /// This is sorted by largest to smallest, so the first element in the list is the largest version.
+  /// Might be `null` if the [result] is not [PubVersionFinderResult.success].
+  final List<Version>? versions;
+
+  /// The result of the version finder.
+  final PubVersionFinderResult? result;
+
+  /// The response object of the http request.
+  final http.Response? httpResponse;
+}
+
+/// An enum representing the result of [PubVersionFinder].
+enum PubVersionFinderResult {
+  /// The version finder successfully found a version.
+  success,
+
+  /// The version finder failed to find a valid version.
+  ///
+  /// This might due to http connection errors or user errors.
+  fail,
+
+  /// The version finder failed to locate the package.
+  ///
+  /// This indicates the package is new.
+  noPackageFound,
 }
 
 /// Finding diffs based on `baseGitDir` and `baseSha`.
@@ -569,7 +679,7 @@ class GitVersionFinder {
   final GitDir baseGitDir;
 
   /// The base sha used to get diff.
-  final String baseSha;
+  final String? baseSha;
 
   static bool _isPubspec(String file) {
     return file.trim().endsWith('pubspec.yaml');
@@ -586,8 +696,7 @@ class GitVersionFinder {
     final io.ProcessResult changedFilesCommand = await baseGitDir
         .runCommand(<String>['diff', '--name-only', baseSha, 'HEAD']);
     print('Determine diff with base sha: $baseSha');
-    final String changedFilesStdout =
-        changedFilesCommand.stdout.toString() ?? '';
+    final String changedFilesStdout = changedFilesCommand.stdout.toString();
     if (changedFilesStdout.isEmpty) {
       return <String>[];
     }
@@ -598,7 +707,8 @@ class GitVersionFinder {
 
   /// Get the package version specified in the pubspec file in `pubspecPath` and
   /// at the revision of `gitRef` (defaulting to the base if not provided).
-  Future<Version> getPackageVersion(String pubspecPath, {String gitRef}) async {
+  Future<Version?> getPackageVersion(String pubspecPath,
+      {String? gitRef}) async {
     final String ref = gitRef ?? (await _getBaseSha());
 
     io.ProcessResult gitShow;
@@ -609,20 +719,19 @@ class GitVersionFinder {
       return null;
     }
     final String fileContent = gitShow.stdout as String;
-    final String versionString = loadYaml(fileContent)['version'] as String;
+    final String? versionString = loadYaml(fileContent)['version'] as String?;
     return versionString == null ? null : Version.parse(versionString);
   }
 
   Future<String> _getBaseSha() async {
-    if (baseSha != null && baseSha.isNotEmpty) {
-      return baseSha;
+    if (baseSha != null && baseSha!.isNotEmpty) {
+      return baseSha!;
     }
 
     io.ProcessResult baseShaFromMergeBase = await baseGitDir.runCommand(
         <String>['merge-base', '--fork-point', 'FETCH_HEAD', 'HEAD'],
         throwOnError: false);
-    if (baseShaFromMergeBase == null ||
-        baseShaFromMergeBase.stderr != null ||
+    if (baseShaFromMergeBase.stderr != null ||
         baseShaFromMergeBase.stdout == null) {
       baseShaFromMergeBase = await baseGitDir
           .runCommand(<String>['merge-base', 'FETCH_HEAD', 'HEAD']);
