@@ -14,8 +14,8 @@ import 'util.dart';
 
 void main() {
   group('test drive_example_command', () {
-    CommandRunner<void> runner;
-    RecordingProcessRunner processRunner;
+    late CommandRunner<void> runner;
+    late RecordingProcessRunner processRunner;
     final String flutterCommand =
         const LocalPlatform().isWindows ? 'flutter.bat' : 'flutter';
     setUp(() {
@@ -55,6 +55,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -100,6 +101,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -143,6 +145,25 @@ void main() {
           throwsA(const TypeMatcher<ToolExit>()));
     });
 
+    test('a plugin without any integration test files is reported as an error',
+        () async {
+      createFakePlugin('plugin',
+          withExtraFiles: <List<String>>[
+            <String>['example', 'lib', 'main.dart'],
+          ],
+          isAndroidPlugin: true,
+          isIosPlugin: true);
+
+      final Directory pluginExampleDirectory =
+          mockPackagesDir.childDirectory('plugin').childDirectory('example');
+
+      createFakePubspec(pluginExampleDirectory, isFlutter: true);
+
+      await expectLater(
+          () => runCapturingPrint(runner, <String>['drive-examples']),
+          throwsA(const TypeMatcher<ToolExit>()));
+    });
+
     test(
         'driving under folder "test_driver" when targets are under "integration_test"',
         () async {
@@ -168,6 +189,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -223,6 +245,8 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
+          'Not supported for the target platform; skipping.',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -255,6 +279,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -300,6 +325,8 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
+          'Not supported for the target platform; skipping.',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -332,6 +359,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -379,6 +407,8 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
+          'Not supported for the target platform; skipping.',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -411,6 +441,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -460,6 +491,8 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
+          'Not supported for the target platform; skipping.',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -492,6 +525,7 @@ void main() {
       expect(
         output,
         orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
           '\n\n',
           'All driver tests successful!',
         ]),
@@ -530,6 +564,29 @@ void main() {
           mockPackagesDir.childDirectory('plugin').childDirectory('example');
 
       createFakePubspec(pluginExampleDirectory, isFlutter: true);
+
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'drive-examples',
+      ]);
+
+      expect(
+        output,
+        orderedEquals(<String>[
+          '\n==========\nChecking plugin...',
+          'Not supported for the target platform; skipping.',
+          '\n\n',
+          'All driver tests successful!',
+        ]),
+      );
+
+      print(processRunner.recordedCalls);
+      // Output should be empty since running drive-examples --macos with no macos
+      // implementation is a no-op.
+      expect(processRunner.recordedCalls, <ProcessCall>[]);
+    });
+
+    test('platform interface plugins are silently skipped', () async {
+      createFakePlugin('aplugin_platform_interface');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
         'drive-examples',
