@@ -4,4 +4,4 @@
 
 import 'package:integration_test/integration_test_driver.dart';
 
-Future<void> main() async => integrationDriver();
+Future<void> main() => integrationDriver();
