@@ -25,13 +25,13 @@ class MyApp extends StatelessWidget {
         // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: _MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const _MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
 class _MyHomePage extends StatefulWidget {
-  _MyHomePage({Key? key, required this.title}) : super(key: key);
+  const _MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -98,7 +98,7 @@ class _MyHomePageState extends State<_MyHomePage> {
           children: <Widget>[
             Text(
               'Button tapped $_counter time${_counter == 1 ? '' : 's'}.',
-              key: ValueKey('CountText'),
+              key: const ValueKey<String>('CountText'),
             ),
           ],
         ),
@@ -106,7 +106,7 @@ class _MyHomePageState extends State<_MyHomePage> {
       floatingActionButton: FloatingActionButton(
         onPressed: _incrementCounter,
         tooltip: 'Increment',
-        child: Icon(Icons.add),
+        child: const Icon(Icons.add),
       ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
