@@ -7,11 +7,11 @@ import 'package:path_provider_platform_interface/path_provider_platform_interfac
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-  PathProviderLinux.register();
+  PathProviderLinux.registerWith();
 
-  setUp(() {});
-
-  tearDown(() {});
+  test('registered instance', () {
+    expect(PathProviderPlatform.instance, isA<PathProviderLinux>());
+  });
 
   test('getTemporaryPath', () async {
     final PathProviderPlatform plugin = PathProviderPlatform.instance;
