@@ -6,8 +6,8 @@
 
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:path_provider_windows/path_provider_windows.dart';
 import 'package:integration_test/integration_test.dart';
+import 'package:path_provider_windows/path_provider_windows.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
