@@ -6,15 +6,16 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider_linux/path_provider_linux.dart';
 import 'package:shared_preferences_linux/shared_preferences_linux.dart';
+import 'package:shared_preferences_platform_interface/shared_preferences_platform_interface.dart';
 
 void main() {
   late MemoryFileSystem fs;
 
+  SharedPreferencesLinux.registerWith();
+
   setUp(() {
     fs = MemoryFileSystem.test();
   });
-
-  tearDown(() {});
 
   Future<String> _getFilePath() async {
     final pathProvider = PathProviderLinux();
@@ -37,6 +38,11 @@ void main() {
     prefs.fs = fs;
     return prefs;
   }
+
+  test('registered instance', () {
+    expect(
+        SharedPreferencesStorePlatform.instance, isA<SharedPreferencesLinux>());
+  });
 
   test('getAll', () async {
     await _writeTestFile('{"key1": "one", "key2": 2}');
