@@ -2,17 +2,22 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart=2.9
+
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io' as io;
 
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:flutter_plugin_tools/src/common.dart';
-import 'package:git/git.dart';
-import 'package:mockito/mockito.dart';
-import 'package:test/test.dart';
 import 'package:flutter_plugin_tools/src/version_check_command.dart';
+import 'package:git/git.dart';
+import 'package:http/http.dart' as http;
+import 'package:http/testing.dart';
+import 'package:mockito/mockito.dart';
 import 'package:pub_semver/pub_semver.dart';
+import 'package:test/test.dart';
 import 'util.dart';
 
 void testAllowedVersion(
@@ -39,20 +44,31 @@ class MockGitDir extends Mock implements GitDir {}
 
 class MockProcessResult extends Mock implements io.ProcessResult {}
 
+const String _redColorMessagePrefix = '\x1B[31m';
+const String _redColorMessagePostfix = '\x1B[0m';
+
+// Some error message was printed in a "Colorized" red message. So `\x1B[31m` and `\x1B[0m` needs to be included.
+String _redColorString(String string) {
+  return '$_redColorMessagePrefix$string$_redColorMessagePostfix';
+}
+
 void main() {
+  const String indentation = '  ';
   group('$VersionCheckCommand', () {
     CommandRunner<void> runner;
     RecordingProcessRunner processRunner;
     List<List<String>> gitDirCommands;
     String gitDiffResponse;
     Map<String, String> gitShowResponses;
+    MockGitDir gitDir;
 
     setUp(() {
       gitDirCommands = <List<String>>[];
       gitDiffResponse = '';
       gitShowResponses = <String, String>{};
-      final MockGitDir gitDir = MockGitDir();
-      when(gitDir.runCommand(any)).thenAnswer((Invocation invocation) {
+      gitDir = MockGitDir();
+      when(gitDir.runCommand(any, throwOnError: anyNamed('throwOnError')))
+          .thenAnswer((Invocation invocation) {
         gitDirCommands.add(invocation.positionalArguments[0] as List<String>);
         final MockProcessResult mockProcessResult = MockProcessResult();
         if (invocation.positionalArguments[0][0] == 'diff') {
@@ -61,6 +77,9 @@ void main() {
         } else if (invocation.positionalArguments[0][0] == 'show') {
           final String response =
               gitShowResponses[invocation.positionalArguments[0][1]];
+          if (response == null) {
+            throw const io.ProcessException('git', <String>['show']);
+          }
           when<String>(mockProcessResult.stdout as String).thenReturn(response);
         } else if (invocation.positionalArguments[0][0] == 'merge-base') {
           when<String>(mockProcessResult.stdout as String).thenReturn('abc123');
@@ -145,6 +164,24 @@ void main() {
       expect(
         output,
         containsAllInOrder(<String>[
+          'No version check errors found!',
+        ]),
+      );
+    });
+
+    test('allows valid version for new package.', () async {
+      createFakePlugin('plugin', includeChangeLog: true, includeVersion: true);
+      gitDiffResponse = 'packages/plugin/pubspec.yaml';
+      gitShowResponses = <String, String>{
+        'HEAD:packages/plugin/pubspec.yaml': 'version: 1.0.0',
+      };
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['version-check']);
+
+      expect(
+        output,
+        containsAllInOrder(<String>[
+          '${indentation}Unable to find pubspec in master. Safe to ignore if the project is new.',
           'No version check errors found!',
         ]),
       );
@@ -301,25 +338,27 @@ void main() {
 * Some changes.
 ''';
       createFakeCHANGELOG(pluginDirectory, changelog);
-      final Future<List<String>> output = runCapturingPrint(
-          runner, <String>['version-check', '--base-sha=master']);
-      await expectLater(
+      bool hasError = false;
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
         output,
-        throwsA(const TypeMatcher<ToolExit>()),
+        containsAllInOrder(<String>[
+          _redColorString('''
+versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
+The version in pubspec.yaml is 1.0.1.
+The first version listed in CHANGELOG.md is 1.0.2.
+'''),
+        ]),
       );
-      try {
-        final List<String> outputValue = await output;
-        await expectLater(
-          outputValue,
-          containsAllInOrder(<String>[
-            '''
-  versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
-  The version in pubspec.yaml is 1.0.1.
-  The first version listed in CHANGELOG.md is 1.0.2.
-  ''',
-          ]),
-        );
-      } on ToolExit catch (_) {}
     });
 
     test('Success if CHANGELOG and pubspec versions match', () async {
@@ -368,25 +407,29 @@ void main() {
 * Some other changes.
 ''';
       createFakeCHANGELOG(pluginDirectory, changelog);
-      final Future<List<String>> output = runCapturingPrint(
-          runner, <String>['version-check', '--base-sha=master']);
-      await expectLater(
+      bool hasError = false;
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
         output,
-        throwsA(const TypeMatcher<ToolExit>()),
-      );
-      try {
-        final List<String> outputValue = await output;
-        await expectLater(
-          outputValue,
-          containsAllInOrder(<String>[
+        containsAllInOrder(<String>[
+          _redColorString(
             '''
-  versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
-  The version in pubspec.yaml is 1.0.0.
-  The first version listed in CHANGELOG.md is 1.0.1.
-  ''',
-          ]),
-        );
-      } on ToolExit catch (_) {}
+versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
+The version in pubspec.yaml is 1.0.0.
+The first version listed in CHANGELOG.md is 1.0.1.
+''',
+          )
+        ]),
+      );
     });
 
     test('Allow NEXT as a placeholder for gathering CHANGELOG entries',
@@ -443,25 +486,28 @@ void main() {
 * Some other changes.
 ''';
       createFakeCHANGELOG(pluginDirectory, changelog);
-      final Future<List<String>> output = runCapturingPrint(
-          runner, <String>['version-check', '--base-sha=master']);
-      await expectLater(
+      bool hasError = false;
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
         output,
-        throwsA(const TypeMatcher<ToolExit>()),
-      );
-      try {
-        final List<String> outputValue = await output;
-        await expectLater(
-          outputValue,
-          containsAllInOrder(<String>[
+        containsAllInOrder(<String>[
+          _redColorString(
             '''
-  versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
-  The version in pubspec.yaml is 1.0.0.
-  The first version listed in CHANGELOG.md is 1.0.1.
-  ''',
-          ]),
-        );
-      } on ToolExit catch (_) {}
+When bumping the version for release, the NEXT section should be incorporated
+into the new version's release notes.
+''',
+          )
+        ]),
+      );
     });
 
     test('Fail if the version changes without replacing NEXT', () async {
@@ -482,25 +528,194 @@ void main() {
 * Some other changes.
 ''';
       createFakeCHANGELOG(pluginDirectory, changelog);
-      final Future<List<String>> output = runCapturingPrint(
-          runner, <String>['version-check', '--base-sha=master']);
-      await expectLater(
+      bool hasError = false;
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
         output,
-        throwsA(const TypeMatcher<ToolExit>()),
-      );
-      try {
-        final List<String> outputValue = await output;
-        await expectLater(
-          outputValue,
-          containsAllInOrder(<String>[
+        containsAllInOrder(<String>[
+          'Found NEXT; validating next version in the CHANGELOG.',
+          _redColorString(
             '''
-  versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
-  The version in pubspec.yaml is 1.0.0.
-  The first version listed in CHANGELOG.md is 1.0.1.
-  ''',
-          ]),
-        );
-      } on ToolExit catch (_) {}
+versions for plugin in CHANGELOG.md and pubspec.yaml do not match.
+The version in pubspec.yaml is 1.0.1.
+The first version listed in CHANGELOG.md is 1.0.0.
+''',
+          )
+        ]),
+      );
+    });
+
+    test('allows valid against pub', () async {
+      const Map<String, dynamic> httpResponse = <String, dynamic>{
+        'name': 'some_package',
+        'versions': <String>[
+          '0.0.1',
+          '0.0.2',
+          '1.0.0',
+        ],
+      };
+      final MockClient mockClient = MockClient((http.Request request) async {
+        return http.Response(json.encode(httpResponse), 200);
+      });
+      final VersionCheckCommand command = VersionCheckCommand(
+          mockPackagesDir, mockFileSystem,
+          processRunner: processRunner, gitDir: gitDir, httpClient: mockClient);
+
+      runner = CommandRunner<void>(
+          'version_check_command', 'Test for $VersionCheckCommand');
+      runner.addCommand(command);
+
+      createFakePlugin('plugin', includeChangeLog: true, includeVersion: true);
+      gitDiffResponse = 'packages/plugin/pubspec.yaml';
+      gitShowResponses = <String, String>{
+        'master:packages/plugin/pubspec.yaml': 'version: 1.0.0',
+        'HEAD:packages/plugin/pubspec.yaml': 'version: 2.0.0',
+      };
+      final List<String> output = await runCapturingPrint(runner,
+          <String>['version-check', '--base-sha=master', '--against-pub']);
+
+      expect(
+        output,
+        containsAllInOrder(<String>[
+          '${indentation}plugin: Current largest version on pub: 1.0.0',
+          'No version check errors found!',
+        ]),
+      );
+    });
+
+    test('denies invalid against pub', () async {
+      const Map<String, dynamic> httpResponse = <String, dynamic>{
+        'name': 'some_package',
+        'versions': <String>[
+          '0.0.1',
+          '0.0.2',
+        ],
+      };
+      final MockClient mockClient = MockClient((http.Request request) async {
+        return http.Response(json.encode(httpResponse), 200);
+      });
+      final VersionCheckCommand command = VersionCheckCommand(
+          mockPackagesDir, mockFileSystem,
+          processRunner: processRunner, gitDir: gitDir, httpClient: mockClient);
+
+      runner = CommandRunner<void>(
+          'version_check_command', 'Test for $VersionCheckCommand');
+      runner.addCommand(command);
+
+      createFakePlugin('plugin', includeChangeLog: true, includeVersion: true);
+      gitDiffResponse = 'packages/plugin/pubspec.yaml';
+      gitShowResponses = <String, String>{
+        'master:packages/plugin/pubspec.yaml': 'version: 1.0.0',
+        'HEAD:packages/plugin/pubspec.yaml': 'version: 2.0.0',
+      };
+
+      bool hasError = false;
+      final List<String> result = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
+        result,
+        containsAllInOrder(<String>[
+          _redColorString(
+            '''
+${indentation}Incorrectly updated version.
+${indentation}HEAD: 2.0.0, pub: 0.0.2.
+${indentation}Allowed versions: {1.0.0: NextVersionType.BREAKING_MAJOR, 0.1.0: NextVersionType.MINOR, 0.0.3: NextVersionType.PATCH}''',
+          )
+        ]),
+      );
+    });
+
+    test(
+        'throw and print error message if http request failed when checking against pub',
+        () async {
+      final MockClient mockClient = MockClient((http.Request request) async {
+        return http.Response('xx', 400);
+      });
+      final VersionCheckCommand command = VersionCheckCommand(
+          mockPackagesDir, mockFileSystem,
+          processRunner: processRunner, gitDir: gitDir, httpClient: mockClient);
+
+      runner = CommandRunner<void>(
+          'version_check_command', 'Test for $VersionCheckCommand');
+      runner.addCommand(command);
+
+      createFakePlugin('plugin', includeChangeLog: true, includeVersion: true);
+      gitDiffResponse = 'packages/plugin/pubspec.yaml';
+      gitShowResponses = <String, String>{
+        'master:packages/plugin/pubspec.yaml': 'version: 1.0.0',
+        'HEAD:packages/plugin/pubspec.yaml': 'version: 2.0.0',
+      };
+      bool hasError = false;
+      final List<String> result = await runCapturingPrint(runner, <String>[
+        'version-check',
+        '--base-sha=master',
+        '--against-pub'
+      ], errorHandler: (Error e) {
+        expect(e, isA<ToolExit>());
+        hasError = true;
+      });
+      expect(hasError, isTrue);
+
+      expect(
+        result,
+        containsAllInOrder(<String>[
+          _redColorString(
+            '''
+${indentation}Error fetching version on pub for plugin.
+${indentation}HTTP Status 400
+${indentation}HTTP response: xx
+''',
+          )
+        ]),
+      );
+    });
+
+    test('when checking against pub, allow any version if http status is 404.',
+        () async {
+      final MockClient mockClient = MockClient((http.Request request) async {
+        return http.Response('xx', 404);
+      });
+      final VersionCheckCommand command = VersionCheckCommand(
+          mockPackagesDir, mockFileSystem,
+          processRunner: processRunner, gitDir: gitDir, httpClient: mockClient);
+
+      runner = CommandRunner<void>(
+          'version_check_command', 'Test for $VersionCheckCommand');
+      runner.addCommand(command);
+
+      createFakePlugin('plugin', includeChangeLog: true, includeVersion: true);
+      gitDiffResponse = 'packages/plugin/pubspec.yaml';
+      gitShowResponses = <String, String>{
+        'master:packages/plugin/pubspec.yaml': 'version: 1.0.0',
+        'HEAD:packages/plugin/pubspec.yaml': 'version: 2.0.0',
+      };
+      final List<String> result = await runCapturingPrint(runner,
+          <String>['version-check', '--base-sha=master', '--against-pub']);
+
+      expect(
+        result,
+        containsAllInOrder(<String>[
+          '${indentation}Unable to find package on pub server. Safe to ignore if the project is new.',
+          'No version check errors found!',
+        ]),
+      );
     });
   });
 
